♻️  리뷰 관련 코드 디렉토리 분리
