+++ src/main/java/com/ssmoker/smoker/domain/smokingArea/dto/SmokingAreaStarInfoResponse.java
@@ -1,0 +1,4 @@
+package com.ssmoker.smoker.domain.smokingArea.dto;
+
+public class SmokingAreaStarInfoResponse {
+}
