+++ src/main/java/com/ssmoker/smoker/domain/smokingArea/repository/SmokingAreaRepository.java
@@ -20,7 +20,7 @@ public interface SmokingAreaRepository extends JpaRepository<SmokingArea, Long> 
             "where s.id = :smokingAreaId")
     int findReviewCountBySmokingAreaId(@Param("smokingAreaId") Long smokingAreaId);
 
-    //null이 아닌 0을 반환하도록 함
+    //null 이 아닌 0을 반환하도록 함
     @Query("select coalesce(count(sa), 0)" +
             "from SmokingArea s left join SavedSmokingArea sa " +
             "on sa.smokingArea.id = s.id " +
