+++ src/main/java/com/ssmoker/smoker/domain/smokingArea/service/SmokingAreaService.java
@@ -29,8 +29,6 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class SmokingAreaService {
 
-    private static final int REVIEW_PAGE_SIZE = 5;
-
     private final SmokingAreaRepository smokingAreaRepository;
     private final ReviewRepository reviewRepository;
 
@@ -42,27 +40,6 @@ public class SmokingAreaService {
         throw new SmokingAreaNotFoundException(SMOKING_AREA_NOT_FOUND);
     }
 
-    public ReviewResponses getReviewsByAreaId(Long id, int pageNumber) {
-        if (pageNumber < 0) {
-            throw new ReviewPageNumberException(REVIEW_BAD_REQUEST);
-        }
-        Page<Review> reviewPage = reviewRepository.findReviewsWithMemberById(id,
-                PageRequest.of(pageNumber, REVIEW_PAGE_SIZE));
-
-        ReviewResponses reviewResponses = ReviewResponses.of(getReviewResponsesByAreaId(reviewPage), reviewPage.isLast(),
-                reviewPage.getNumber());
-        return reviewResponses;
-    }
-
-    private List<ReviewResponse> getReviewResponsesByAreaId(Page<Review> reviewPage) {
-        return reviewPage.getContent().stream()
-                .map(this::getReviewResponse)
-                .collect(Collectors.toList());
-    }
-
-    private ReviewResponse getReviewResponse(Review review) {
-        return ReviewResponse.of(review, review.getMember().getNickName());
-    }
 
     //marker를 위한 모든 db 보내기
     public MapResponse.SmokingMarkersResponse getSmokingMarkersResponse() {
